chore(tuning):tuning, vertx

+++ src/main/java/dto/PaymentSummary.java
@@ -1,9 +1,7 @@
 package dto;
 
-import java.math.BigDecimal;
-
 public record PaymentSummary(
         int totalRequests,
-        BigDecimal totalAmount
+        double totalAmount
 ) {
 }
